+++ install.php
@@ -3,5 +3,4 @@
 if(!defined('SIMWP_LOADED')){
 	require __DIR__ . '/vendor/autoload.php';
 	require __DIR__ . '/Simwp.php';
-	define('SIMWP_LOADED', true);
 }