+++ src/Component/Level.php
@@ -1,0 +1,16 @@
+<?php
+namespace Simwp\Component;
+use Simwp;
+
+class Level extends Section {
+	public function appendTo($options){
+		foreach($options as $option){
+			if($option instanceof Option){
+				$option->level = $this->name;
+			}
+			else{
+				Simwp::option($option)->level = $this->name;
+			}
+		}
+	}
+}