+++ src/Form/DatePicker.php
@@ -6,7 +6,8 @@ class DatePicker extends Input {
 	public function __construct(){
 		wp_enqueue_script('jquery-ui-core');
 		wp_enqueue_script('jquery-ui-datepicker');
-		if($locale = get_locale() != 'en_US' && $locale){
+		$locale = get_locale();
+		if($locale && $locale != 'en_US'){
 			wp_enqueue_script('jquery-ui-datepicker/locale' , Simwp::url( Simwp::PATH . '/extras/jquery-ui/datepicker-locales/datepicker-' . $locale . '.js'), array('jquery-ui-datepicker'), false, true);
 		}
 	}